Add files via upload
Have a base model with the player able to move with "a" and "d" keys, an is able to jump with "w". When the player eraches the side of a screen, he is transitioned to the next area. The loop goes forrestL>forrestM>forrestR>plainsL>plainsM>plainsR>cliffL>cliff>M>cliffR